convert code to TypeScript, see https://github.com/phetsims/isotopes-and-atomic-mass/issues/115

+++ js/make-isotopes/model/MakeIsotopesModel.ts
@@ -10,7 +10,7 @@
  * @author Aadish Gupta
  */
 
-import createObservableArray from '../../../../axon/js/createObservableArray.js';
+import createObservableArray, { ObservableArray } from '../../../../axon/js/createObservableArray.js';
 import Emitter from '../../../../axon/js/Emitter.js';
 import Dimension2 from '../../../../dot/js/Dimension2.js';
 import Vector2 from '../../../../dot/js/Vector2.js';
@@ -25,55 +25,58 @@ import Tandem from '../../../../tandem/js/Tandem.js';
 import isotopesAndAtomicMass from '../../isotopesAndAtomicMass.js';
 import IsotopesAndAtomicMassStrings from '../../IsotopesAndAtomicMassStrings.js';
 
-const neutronsString = IsotopesAndAtomicMassStrings.neutrons;
+type IsDraggingListener = ( isDragging: boolean ) => void;
 
-// constants
+const neutronsString: string = IsotopesAndAtomicMassStrings.neutrons;
+
 const DEFAULT_NUM_NEUTRONS_IN_BUCKET = 4;
 const NUCLEUS_JUMP_PERIOD = 0.1; // In seconds
 const MAX_NUCLEUS_JUMP = ShredConstants.NUCLEON_RADIUS * 0.5;
-const JUMP_ANGLES = [ Math.PI * 0.1, Math.PI * 1.6, Math.PI * 0.7, Math.PI * 1.1, Math.PI * 0.3 ];
-const JUMP_DISTANCES = [ MAX_NUCLEUS_JUMP * 0.4, MAX_NUCLEUS_JUMP * 0.8, MAX_NUCLEUS_JUMP * 0.2, MAX_NUCLEUS_JUMP * 0.9 ];
-const NUCLEON_CAPTURE_RADIUS = 100; // maximum drop distance for a nucleon to be considered part of the particle
+const JUMP_ANGLES: number[] = [ Math.PI * 0.1, Math.PI * 1.6, Math.PI * 0.7, Math.PI * 1.1, Math.PI * 0.3 ];
+const JUMP_DISTANCES: number[] = [ MAX_NUCLEUS_JUMP * 0.4, MAX_NUCLEUS_JUMP * 0.8, MAX_NUCLEUS_JUMP * 0.2, MAX_NUCLEUS_JUMP * 0.9 ];
+const NUCLEON_CAPTURE_RADIUS = 100;
 const BUCKET_SIZE = new Dimension2( 130, 60 );
 const NEUTRON_BUCKET_POSITION = new Vector2( -220, -180 );
-const DEFAULT_ATOM_CONFIG = new NumberAtom( { protonCount: 1, neutronCount: 0, electronCount: 1 } ); // Hydrogen.
+const DEFAULT_ATOM_CONFIG = new NumberAtom( { protonCount: 1, neutronCount: 0, electronCount: 1 } ); // Hydrogen atom
 
 class MakeIsotopesModel {
 
-  /**
-   * Constructor for a make isotopes model.  This will construct the model with atoms initially in the bucket.
-   *
-   */
-  constructor() {
+  public readonly particleAtom: ParticleAtom;
+  public readonly numberAtom: NumberAtom;
+  public readonly atomReconfigured: Emitter;
+  public nucleusStable: boolean;
+  private nucleusJumpCountdown: number;
+  private nucleusJumpCount: number;
+  public readonly neutrons: ObservableArray<Particle>;
+  public readonly protons: ObservableArray<Particle>;
+  public readonly electrons: ObservableArray<Particle>;
+  public readonly neutronBucket: SphereBucket<Particle>;
+  private readonly mapNeutronsToDragListeners = new Map<Particle, IsDraggingListener>();
 
-    // @public - create the atom.
+  public constructor() {
     this.particleAtom = new ParticleAtom( {
-      tandem: Tandem.OPT_OUT // Opt out for now until phet-io is implemented.
+      tandem: Tandem.OPT_OUT
     } );
 
-    // @public - Make available a 'number atom' that tracks the state of the particle atom.
     this.numberAtom = new NumberAtom( {
       protonCount: DEFAULT_ATOM_CONFIG.protonCountProperty.get(),
       neutronCount: DEFAULT_ATOM_CONFIG.neutronCountProperty.get(),
       electronCount: DEFAULT_ATOM_CONFIG.electronCountProperty.get()
     } );
 
-    // @public - events emitted by instances of this type
     this.atomReconfigured = new Emitter();
 
-    // Update the stability state and counter on changes.
-    this.nucleusStable = true; // @public
-    this.nucleusJumpCountdown = NUCLEUS_JUMP_PERIOD; // @private
-    this.nucleusOffset = Vector2.ZERO; // @private
-    this.nucleusJumpCount = 0; // @private
+    this.nucleusStable = true;
+    this.nucleusJumpCountdown = NUCLEUS_JUMP_PERIOD;
+    this.nucleusJumpCount = 0;
 
-    // Unlink in not required here as it is used through out the sim life
-    this.particleAtom.massNumberProperty.link( massNumber => {
+    this.particleAtom.massNumberProperty.link( ( massNumber: number ) => {
       const stable = massNumber > 0 ?
-                     AtomIdentifier.isStable( this.particleAtom.protonCountProperty.get(),
-                       this.particleAtom.neutronCountProperty.get() ) : true;
+                     AtomIdentifier.isStable(
+                       this.particleAtom.protonCountProperty.get(),
+                       this.particleAtom.neutronCountProperty.get()
+                     ) : true;
       if ( this.nucleusStable !== stable ) {
-        // Stability has changed.
         this.nucleusStable = stable;
         if ( stable ) {
           this.nucleusJumpCountdown = NUCLEUS_JUMP_PERIOD;
@@ -85,15 +88,10 @@ class MakeIsotopesModel {
       }
     } );
 
-    // Arrays that contain the subatomic particles, whether they are in the  bucket or in the atom.  This is part of a
-    // basic assumption about how the model works, which is that the model contains all the particles, and the particles
-    // move back and forth from being in the bucket or in in the atom.
-    this.neutrons = createObservableArray(); // @public
-    this.protons = createObservableArray(); // @public
-    this.electrons = createObservableArray(); // @public
+    this.neutrons = createObservableArray();
+    this.protons = createObservableArray();
+    this.electrons = createObservableArray();
 
-    // The bucket that holds the neutrons that are not in the atom.
-    // @public
     this.neutronBucket = new SphereBucket( {
       position: NEUTRON_BUCKET_POSITION,
       size: BUCKET_SIZE,
@@ -106,24 +104,19 @@ class MakeIsotopesModel {
       this.setAtomConfiguration( this.numberAtom );
     } );
 
-    // Set the initial atom configuration.
     this.setAtomConfiguration( DEFAULT_ATOM_CONFIG );
-
   }
 
-  // Main model step function, called by the framework.
-  // @public
-  step( dt ) {
-    // Update particle positions.
+  public step( dt: number ): void {
     this.neutrons.forEach( neutron => {
       neutron.step( dt );
     } );
 
-    this.protons.forEach( neutron => {
-      neutron.step( dt );
+    this.protons.forEach( proton => {
+      proton.step( dt );
     } );
 
-    if ( this.nucleusStable === false ) {
+    if ( !this.nucleusStable ) {
       this.nucleusJumpCountdown -= dt;
       if ( this.nucleusJumpCountdown <= 0 ) {
         this.nucleusJumpCountdown = NUCLEUS_JUMP_PERIOD;
@@ -132,7 +125,8 @@ class MakeIsotopesModel {
           const angle = JUMP_ANGLES[ this.nucleusJumpCount % JUMP_ANGLES.length ];
           const distance = JUMP_DISTANCES[ this.nucleusJumpCount % JUMP_DISTANCES.length ];
           this.particleAtom.nucleusOffsetProperty.set(
-            new Vector2( Math.cos( angle ) * distance, Math.sin( angle ) * distance ) );
+            new Vector2( Math.cos( angle ) * distance, Math.sin( angle ) * distance )
+          );
         }
         else {
           this.particleAtom.nucleusOffsetProperty.set( Vector2.ZERO );
@@ -141,49 +135,49 @@ class MakeIsotopesModel {
     }
   }
 
-  /**
-   * Get the current atom of this model, in its number representation
-   * @public
-   */
-  getNumberAtom() {
+  public getNumberAtom(): NumberAtom {
     return this.numberAtom;
   }
 
   /**
-   *
-   * @param {Particle} particle
-   * @param {SphereBucket} bucket
-   * @param {ParticleAtom} atom
-   * @public
+   * Places a neutron in the particle atom if it is close enough, otherwise places it in the neutron bucket.
    */
-  placeNucleon( particle, bucket, atom ) {
-    if ( particle.positionProperty.get().distance( atom.positionProperty.get() ) < NUCLEON_CAPTURE_RADIUS ) {
-      atom.addParticle( particle );
+  private placeNucleon( particle: Particle ): void {
+    assert && assert( particle.type === 'neutron', 'Only neutrons should be interactive in the make isotopes model.' );
+    if ( particle.positionProperty.get().distance( this.particleAtom.positionProperty.get() ) < NUCLEON_CAPTURE_RADIUS ) {
+      this.particleAtom.addParticle( particle );
     }
     else {
-      bucket.addParticleNearestOpen( particle, true );
+      this.neutronBucket.addParticleNearestOpen( particle, true );
     }
   }
 
   /**
-   *
-   * @param {Particle} neutron
-   * @param {boolean} lazyLink whether the linking has to be lazy or not
-   * @private
+   * Add a drag listener to the neutron, which will handle moving the neutron between the atom and the neutron bucket.
    */
-  linkNeutron( neutron, lazyLink ) {
-    const isDraggingHandler = isDragging => {
+  private addDragListener( neutron: Particle, lazyLink: boolean ): void {
+
+    assert && assert( !this.mapNeutronsToDragListeners.has( neutron ), 'Neutron already has a drag listener linked.' );
+
+    const isDraggingHandler = ( isDragging: boolean ) => {
       if ( isDragging ) {
         if ( neutron.containerProperty.value ) {
 
+          const preRemovalContainer = neutron.containerProperty.value;
+
           // Remove the neutron from its container, which will be either a bucket or the particle atom.
           neutron.containerProperty.value.removeParticle( neutron );
-          this.atomReconfigured.emit();
+
+          if ( preRemovalContainer === this.particleAtom ) {
+            this.atomReconfigured.emit();
+          }
         }
       }
-      else if ( !isDragging && !this.neutronBucket.includes( neutron ) ) {
-        this.placeNucleon( neutron, this.neutronBucket, this.particleAtom );
-        this.atomReconfigured.emit();
+      else if ( !isDragging && neutron.containerProperty.value === null ) {
+        this.placeNucleon( neutron );
+        if ( neutron.containerProperty.value === this.particleAtom ) {
+          this.atomReconfigured.emit();
+        }
       }
     };
     if ( lazyLink ) {
@@ -192,48 +186,41 @@ class MakeIsotopesModel {
     else {
       neutron.isDraggingProperty.link( isDraggingHandler );
     }
-    neutron.isDraggingHandlerUnlink = () => {
-      neutron.isDraggingProperty.unlink( isDraggingHandler );
-    };
-  }
-
-  // @public
-  setNeutronBucketConfiguration() {
-
-    // Add the neutrons to the neutron bucket.
-    _.times( DEFAULT_NUM_NEUTRONS_IN_BUCKET, () => {
-      const neutron = new Particle( 'neutron' );
-      this.neutronBucket.addParticleFirstOpen( neutron, false );
-      this.linkNeutron( neutron, false );
-      this.neutrons.add( neutron );
-    } );
+    this.mapNeutronsToDragListeners.set( neutron, isDraggingHandler );
   }
 
   /**
-   * Set the configuration of the atom that the user interacts with.  Specifically, this sets the particle atom equal
-   * to the current number atom.  This is done here rather than by directly accessing the atom so that the
-   * appropriate notifications can be sent and the bucket can be
-   * reinitialized.
-   *
-   * @param {NumberAtom} numberAtom - New configuration of atomic properties to which the atom should be set.
-   * @public
+   * Removes the drag listener for the neutron to avoid memory leaks.
    */
-  setAtomConfiguration( numberAtom ) {
+  private removeDragListener( neutron: Particle ): void {
+    assert && assert( this.mapNeutronsToDragListeners.has( neutron ), 'Neutron does not have a drag listener linked.' );
+
+    // Remove the drag listener for the neutron.
+    const isDraggingHandler = this.mapNeutronsToDragListeners.get( neutron );
+    if ( isDraggingHandler ) {
+      neutron.isDraggingProperty.unlink( isDraggingHandler );
+      this.mapNeutronsToDragListeners.delete( neutron );
+    }
+  }
+
+  public setAtomConfiguration( numberAtom: NumberAtom ): void {
+
+    // Clear the current atom configuration.
     this.particleAtom.clear();
     this.protons.clear();
     this.electrons.clear();
-    this.neutrons.forEach( neutron => {
-      neutron.isDraggingHandlerUnlink();
-    } );
+    this.neutrons.forEach( neutron => { this.removeDragListener( neutron ); } );
     this.neutrons.clear();
     this.neutronBucket.reset();
+
+    // Set the new atom configuration.
     if ( this.numberAtom !== numberAtom ) {
       this.numberAtom.protonCountProperty.set( numberAtom.protonCountProperty.get() );
       this.numberAtom.electronCountProperty.set( numberAtom.electronCountProperty.get() );
       this.numberAtom.neutronCountProperty.set( numberAtom.neutronCountProperty.get() );
     }
 
-    // Add the particles.
+    // Create the particles for the atom based on the number atom's properties.
     for ( let i = 0; i < numberAtom.electronCountProperty.get(); i++ ) {
       const electron = new Particle( 'electron' );
       this.particleAtom.addParticle( electron );
@@ -244,35 +231,28 @@ class MakeIsotopesModel {
       this.particleAtom.addParticle( proton );
       this.protons.add( proton );
     }
-    _.times( numberAtom.neutronCountProperty.get(), () => {
+    for ( let k = 0; k < numberAtom.neutronCountProperty.get(); k++ ) {
       const neutron = new Particle( 'neutron' );
       this.particleAtom.addParticle( neutron );
       this.neutrons.add( neutron );
-      this.linkNeutron( neutron, true );
-    } );
+      this.addDragListener( neutron, true );
+    }
     this.particleAtom.moveAllParticlesToDestination();
-    this.setNeutronBucketConfiguration();
+
+    // Add the neutrons to the neutron bucket.
+    _.times( DEFAULT_NUM_NEUTRONS_IN_BUCKET, () => {
+      const neutron = new Particle( 'neutron' );
+      this.neutronBucket.addParticleFirstOpen( neutron, false );
+      this.addDragListener( neutron, false );
+      this.neutrons.add( neutron );
+    } );
+
+    // Emit an event to notify that the atom has been reconfigured.
     this.atomReconfigured.emit();
   }
 
-  /**
-   * Reset the model. The sets the atom and the neutron bucket into their default initial states.
-   * @public
-   */
-  reset() {
-    // Reset the atom.  This also resets the neutron bucket.
+  public reset(): void {
     this.setAtomConfiguration( DEFAULT_ATOM_CONFIG );
-  }
-
-  /**
-   * Get neutron bucket.
-   *
-   * @returns { SphereBucket }
-   *
-   * @public
-   */
-  getNeutronBucket() {
-    return this.neutronBucket;
   }
 }
